+++ src/pages/api/improve-blogs.js
@@ -13,35 +13,7 @@ export default async function handler(req, res) {
             try {
                 res.status(200).json({status:"success"})   // send immediate res 
                 const blogs = await blogServices.getLastHourBlogs(environment);
-                const bulkOperations = await Promise.all(blogs.map(async (blog) => {
-                    try{
-                    const countryCode = blog.countryCode || 'IN'; 
-                    const auther =await getNames(countryCode);
-                    let aiResponse = await askAi(
-                      process.env.IMPROVE_BRIDGE,
-                      `${improveBlogPrompt( auther.writer.name, auther.philosopher.name , countryCode )} ${JSON.stringify({blog : blog.blog , title : blog.title })}`
-                    );
-                    const message_id = aiResponse.response.data.message_id;
-                    aiResponse = extractJsonFromMarkdown(aiResponse.response.data.content);
-                    const processedBlog = ValidateAiResponse(aiResponse, improveBlogSchema,process.env.IMPROVE_BRIDGE,message_id,true);
-                    await distinctifyPhrase(processedBlog, environment);
-                    return {
-                        updateOne: {
-                            filter: { id: blog.id },
-                            update: { 
-                              $set: { 
-                                'blog': processedBlog.blog ,
-                                'title' : processedBlog.title,
-                              }
-                            }
-                        }
-                    };
-                }catch(err){
-                    console.log(err,"error in ask ai ")
-                    sendMessageTochannel({"message":'error in improve blog askAi.' , error : err.message})
-                    return null ;
-                }
-            }));
+                const bulkOperations =await createBulkOperation(blogs,environment);
                 const validBulkOperations = bulkOperations.filter(op => op !== null);            
                 results = await blogServices.bulkUpdateBlogs(validBulkOperations, environment);
             } catch (error) {
@@ -57,7 +29,7 @@ export default async function handler(req, res) {
     }
 }
 
-async function distinctifyPhrase(processedBlog, environment) {
+export async function distinctifyPhrase(processedBlog, environment) {
     if(processedBlog?.phrase){
         const existingBlogs = await blogServices.searchBlogsByQuery(processedBlog.phrase.content, environment);
         if(existingBlogs.length <= 5) return;
@@ -68,7 +40,7 @@ async function distinctifyPhrase(processedBlog, environment) {
     }
 }
 
-async function getNames (countryCode){
+export async function getNames (countryCode){
     let [writer, philosopher] = await Promise.all([
         getRandomAuthorByCountryAndType(countryCode, 'writer'),
         getRandomAuthorByCountryAndType(countryCode, 'philosopher')
@@ -82,7 +54,7 @@ async function getNames (countryCode){
     }
     return { writer , philosopher }
 }
-async function getWriter(countryCode,type){
+export async function getWriter(countryCode,type){
     const res = await askAi(process.env.NEXT_PUBLIC_WRITER_GENERATOR_BRIGDE,`countryCode:${countryCode} and type:${type}`,{ countryCode : countryCode , profession : type });
     const names = JSON.parse(res.response.data.content).data;
     try {
@@ -91,4 +63,37 @@ async function getWriter(countryCode,type){
         console.log(error)
     }
     return names[Math.floor(Math.random() * names.length)];
+}
+
+
+export async function createBulkOperation (blogs,environment){
+   return await Promise.all(blogs.map(async (blog) => {
+        try{
+        const countryCode = blog.countryCode || 'IN'; 
+        const auther =await getNames(countryCode);
+        let aiResponse = await askAi(
+          process.env.IMPROVE_BRIDGE,
+          `${improveBlogPrompt( auther.writer.name, auther.philosopher.name , countryCode )} ${JSON.stringify({blog : blog.blog , title : blog.title })}`
+        );
+        const message_id = aiResponse.response.data.message_id;
+        aiResponse = extractJsonFromMarkdown(aiResponse.response.data.content);
+        const processedBlog = ValidateAiResponse(aiResponse, improveBlogSchema,process.env.IMPROVE_BRIDGE,message_id,true);
+        await distinctifyPhrase(processedBlog, environment);
+        return {
+            updateOne: {
+                filter: { id: blog.id },
+                update: { 
+                  $set: { 
+                    'blog': processedBlog.blog ,
+                    'title' : processedBlog.title,
+                  }
+                }
+            }
+        };
+    }catch(err){
+        console.log(err,"error in ask ai ")
+        sendMessageTochannel({"message":'error in improve blog askAi.' , error : err.message})
+        return null ;
+    }
+}));
 }
